+++ niftilib/nifti1_tool.c
@@ -760,7 +760,7 @@ int add_int(int_list * ilist, int val)
 {
    if( ilist->len == 0 ) ilist->list = NULL;  /* just to be safe */
    ilist->len++;
-   ilist->list = (int *)realloc(ilist->list,ilist->len*sizeof(int));
+   ilist->list = (int *)realloc(ilist->list,(size_t)ilist->len * sizeof(int));
    if( ! ilist->list ){
       fprintf(stderr,"** failed to alloc %d (int *) elements\n",ilist->len);
       return -1;
@@ -781,7 +781,7 @@ int add_string(str_list * slist, const char * str)
 {
    if( slist->len == 0 ) slist->list = NULL;  /* just to be safe */
    slist->len++;
-   slist->list = (const char **)realloc(slist->list,slist->len*sizeof(char *));
+   slist->list = (const char **)realloc(slist->list,(size_t)slist->len * sizeof(char *));
    if( ! slist->list ){
       fprintf(stderr,"** failed to alloc %d (char *) elements\n",slist->len);
       return -1;
@@ -1935,14 +1935,14 @@ static char * read_file_text(const char * filename, int * length)
 
    /* allocate the bytes, and fill them with the file contents */
 
-   text = (char *)malloc(len * sizeof(char));
+   text = (char *)malloc((size_t)len * sizeof(char));
    if( !text ) {
       fprintf(stderr,"** RFT: failed to allocate %d bytes\n", len);
       fclose(fp);
       return NULL;
    }
 
-   bytes = fread(text, sizeof(char), len, fp);
+   bytes = fread(text, sizeof(char),(size_t)len, fp);
    fclose(fp); /* in any case */
 
    if( bytes != (size_t)len ) {
@@ -2140,7 +2140,7 @@ int remove_ext_list( nifti_image * nim, const char ** elist, int len )
    if( g_debug > 2 )
       fprintf(stderr,"+d removing %d exts from '%s'\n", len, nim->fname );
 
-   if( ! (marks = (int *)calloc(nim->num_ext, sizeof(int))) ) {
+   if( ! (marks = (int *)calloc((size_t)(nim->num_ext),(size_t)sizeof(int))) ) {
       fprintf(stderr,"** failed to alloc %d marks\n",nim->num_ext);
       return -1;
    }
@@ -2972,7 +2972,7 @@ int modify_field(void * basep, field_s * field, const char * data)
                }
                /* otherwise, we're good */
                { const int16_t sval = (int16_t)val;
-                  memcpy((char *)basep + field->offset + fc*sizeof(sval), &sval, sizeof(sval)); }
+                  memcpy((char *)basep + field->offset + (size_t)fc * sizeof(sval), &sval,(size_t)sizeof(sval)); }
                if( g_debug > 1 )
                   fprintf(stderr,"+d setting posn %d of '%s' to %d\n",
                           fc, field->name, val);
@@ -2992,7 +2992,7 @@ int modify_field(void * basep, field_s * field, const char * data)
                   return 1;
                }
                { const int32_t ival = (int32_t)val;
-                  memcpy((char *)basep + field->offset + fc*sizeof(ival), &ival, sizeof(ival)); }
+                  memcpy((char *)basep + field->offset + (size_t)fc * sizeof(ival), &ival,(size_t)sizeof(ival)); }
                if( g_debug > 1 )
                   fprintf(stderr,"+d setting posn %d of '%s' to %d\n",
                           fc, field->name, val);
@@ -3012,7 +3012,7 @@ int modify_field(void * basep, field_s * field, const char * data)
                   return 1;
                }
                /* otherwise, we're good */
-               memcpy((char *)basep + field->offset + fc*sizeof(fval), &fval, sizeof(fval));
+               memcpy((char *)basep + field->offset + (size_t)fc * sizeof(fval), &fval,(size_t)sizeof(fval));
                if( g_debug > 1 )
                   fprintf(stderr,"+d setting posn %d of '%s' to %f\n",
                           fc, field->name, fval);
@@ -3025,9 +3025,9 @@ int modify_field(void * basep, field_s * field, const char * data)
          {
             char * dest = (char *)basep + field->offset;
             nchars = (int)dataLength;
-            strncpy(dest, data, field->len);
+            strncpy(dest, data,(size_t)(field->len));
             if( nchars < field->len )  /* clear the rest */
-               memset(dest+nchars, '\0', field->len-nchars);
+               memset(dest+nchars, '\0',(size_t)(field->len-nchars));
          }
          break;
    }
@@ -4266,7 +4266,7 @@ nifti_image * nt_read_bricks(nt_opts * opts, const char * fname, int len, int * 
     /* now populate NBL (can be based only on len and nim) */
     NBL->nbricks = len;
     NBL->bsize = (size_t)nim->nbyper * nim->nx * nim->ny * nim->nz;
-    NBL->bricks = (void **)calloc(NBL->nbricks, sizeof(void *));
+    NBL->bricks = (void **)calloc((size_t)(NBL->nbricks),(size_t)(sizeof(void *)));
     if( !NBL->bricks ){
         fprintf(stderr,"** NRB: failed to alloc %d pointers\n",NBL->nbricks);
         nifti_image_free(nim);
